借助AppLog表存储OAuthServer相关登录信息

+++ NewLife.Cube/Entity/应用日志.Biz.cs
@@ -59,7 +59,7 @@ namespace NewLife.Cube.Entity
         /// <summary>根据编号查找</summary>
         /// <param name="id">编号</param>
         /// <returns>实体对象</returns>
-        public static AppLog FindByID(Int32 id)
+        public static AppLog FindByID(Int64 id)
         {
             if (id <= 0) return null;
